+++ modelstudio/src/com/hiveworkshop/rms/ui/application/edit/mesh/types/tpose/TPoseSelectionManager.java
@@ -41,9 +41,9 @@ public final class TPoseSelectionManager extends AbstractSelectionManager<IdObje
 		centerOfGroupSumHeap.z = 0;
 		for (final IdObject object : selection) {
 			final Vec3 pivot = object.getPivotPoint();
-			pivotHeap.x = (float) pivot.x;
-			pivotHeap.y = (float) pivot.y;
-			pivotHeap.z = (float) pivot.z;
+			pivotHeap.x = pivot.x;
+			pivotHeap.y = pivot.y;
+			pivotHeap.z = pivot.z;
 			centerOfGroupSumHeap.x += pivotHeap.x;
 			centerOfGroupSumHeap.y += pivotHeap.y;
 			centerOfGroupSumHeap.z += pivotHeap.z;
@@ -61,9 +61,9 @@ public final class TPoseSelectionManager extends AbstractSelectionManager<IdObje
 		double radius = 0;
 		for (final IdObject item : selection) {
 			final Vec3 pivot = item.getPivotPoint();
-			pivotHeap.x = (float) pivot.x;
-			pivotHeap.y = (float) pivot.y;
-			pivotHeap.z = (float) pivot.z;
+			pivotHeap.x = pivot.x;
+			pivotHeap.y = pivot.y;
+			pivotHeap.z = pivot.z;
 			pivotHeap.w = 1;
 			final double distance = sphereCenter.distance(pivotHeap);
 			if (distance >= radius) {
